refactor: Move `Faker::BossaNova` under `Faker::Music::BossaNova`
(fixes #3139)

Moves +Faker::BossaNova+ generator under +Faker::Music::BossaNova+
