--- doc/default/bossa_nova.md
@@ -1,7 +1,0 @@
-# Faker::BossaNova
-
-```ruby
-Faker::BossaNova.artist #=> "Tom Jobim"
-
-Faker::BossaNova.song #=> "Chega de Saudade"
-```
