+++ doc/music/bossa_nova.md
@@ -1,0 +1,7 @@
+# Faker::Music::BossaNova
+
+```ruby
+Faker::Music::BossaNova.artist #=> "Tom Jobim"
+
+Faker::Music::BossaNova.song #=> "Chega de Saudade"
+```
